definition provider code cleanup
